refactor: Refactor bucket context http middleware
